# Bm25: only add additonal explanations when requested