--- a/adapters/handlers/graphql/local/common_filters/bm25.go
+++ b/adapters/handlers/graphql/local/common_filters/bm25.go
@@ -31,6 +31,13 @@ func ExtractBM25(source map[string]interface{}) searchparams.KeywordRanking {
 		args.Query = query.(string)
 	}
 
+	addExplanations, ok := source["additionalExplanations"]
+	if ok {
+		args.AdditionalExplanations = addExplanations.(bool)
+	} else {
+		args.AdditionalExplanations = false
+	}
+
 	args.Type = "bm25"
 
 	return args