--- a/adapters/handlers/graphql/local/get/class_builder_fields.go
+++ b/adapters/handlers/graphql/local/get/class_builder_fields.go
@@ -369,11 +369,10 @@ func (r *resolver) makeResolveGetClass(className string) graphql.FieldResolveFn 
 			}
 		}
 
-		var keywordRankingParams *searchparams.KeywordRanking
-
 		// extracts bm25 (sparseSearch) from the query
+		var keywordRankingParams *searchparams.KeywordRanking
 		if bm25, ok := p.Args["bm25"]; ok {
-			p := common_filters.ExtractBM25(bm25.(map[string]interface{}))
+			p := common_filters.ExtractBM25(bm25.(map[string]interface{}), additional.ExplainScore)
 			keywordRankingParams = &p
 		}
 
@@ -382,7 +381,7 @@ func (r *resolver) makeResolveGetClass(className string) graphql.FieldResolveFn 
 		// refactored
 		var hybridParams *searchparams.HybridSearch
 		if hybrid, ok := p.Args["hybrid"]; ok {
-			p, err := common_filters.ExtractHybridSearch(hybrid.(map[string]interface{}))
+			p, err := common_filters.ExtractHybridSearch(hybrid.(map[string]interface{}), additional.ExplainScore)
 			if err != nil {
 				return nil, fmt.Errorf("failed to extract hybrid params: %w", err)
 			}