--- a/adapters/handlers/graphql/local/get/sparse_search.go
+++ b/adapters/handlers/graphql/local/get/sparse_search.go
@@ -39,5 +39,9 @@ func bm25Fields(prefix string) graphql.InputObjectConfigFieldMap {
 			Description: "The properties to search in",
 			Type:        graphql.NewList(graphql.String),
 		},
+		"additionalExplanations": &graphql.InputObjectFieldConfig{
+			Description: "Return additional explanations for the BM25 results. False by default",
+			Type:        graphql.Boolean,
+		},
 	}
 }