### What's being changed:

Skip some expensive computations if "explainScore" is not part of `_additional{}`.

-----Benchmark-----

old:
```
Min:    1.856917ms
Mean: 556.972387ms
Max:  1.408237916s
p50:  651.758666ms
p90:  1.025386916s
p99:  1.387304041s
```

new:
```
Min:       3.473ms
Mean: 217.437607ms
Max:  930.479709ms
p50:     234.578ms
p90:  376.472667ms
p99:  834.601375ms
```

msmarco first 200 queries

### Review checklist

- [x] Documentation has been updated, if necessary. Link to changed documentation:
- [x] All new code is covered by tests where it is reasonable.
- [x] Performance tests have been run or not necessary.