@byronvoorbach 2023-03-06
This is a great addition @dirkkul!
I would change the name from `additionalExplanations` to `explain` and also add support for Hybrid